requirements.txt
Flask>=3.1.0
Flask-Login>=0.6.3
Flask-SQLAlchemy>=3.1.1
Jinja2>=3.1.5
requests>=2.32.3
SQLAlchemy>=2.0.37
python-dotenv>=1.0.1
Flask-RESTful>=0.3.10
ping3>=4.0.8
